Propose adding TagEveryone_TheBot to README
Propose to add @TagEveryone_TheBot to the list. TagEveryoneTelegramBot is a Telegram bot that lets users mention all group members with /everyone﻿ or @all﻿, similar to Discord mentions. Members opt in via /in﻿, but new users are now added automatically. It supports WebApp integration, Docker deployment, and stores data in SQLite3 for easy self-hosting.